feat: implemented user validation
